Extract jgit from core to its own module
Fix spotless problems with some classes
