+++ konf-core/src/main/kotlin/com/uchuhimo/konf/source/base/ValueSource.kt
@@ -19,11 +19,25 @@ package com.uchuhimo.konf.source.base
 import com.uchuhimo.konf.Path
 import com.uchuhimo.konf.SizeInBytes
 import com.uchuhimo.konf.notEmptyOr
-import com.uchuhimo.konf.source.*
+import com.uchuhimo.konf.source.NoSuchPathException
+import com.uchuhimo.konf.source.Source
+import com.uchuhimo.konf.source.SourceInfo
+import com.uchuhimo.konf.source.WrongTypeException
+import com.uchuhimo.konf.source.toDescription
 import java.math.BigDecimal
 import java.math.BigInteger
-import java.time.*
-import java.util.*
+import java.time.Duration
+import java.time.Instant
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.time.LocalTime
+import java.time.OffsetDateTime
+import java.time.OffsetTime
+import java.time.Year
+import java.time.YearMonth
+import java.time.ZoneOffset
+import java.time.ZonedDateTime
+import java.util.Date
 
 /**
  * Source from a single value.
