+++ konf-core/src/test/kotlin/com/uchuhimo/konf/source/DefaultLoadersSpec.kt
@@ -26,7 +26,6 @@ import com.uchuhimo.konf.tempFileOf
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
-import org.eclipse.jgit.api.Git
 import org.jetbrains.spek.api.dsl.given
 import org.jetbrains.spek.api.dsl.it
 import org.jetbrains.spek.api.dsl.on
@@ -34,7 +33,6 @@ import org.jetbrains.spek.subject.SubjectSpek
 import org.jetbrains.spek.subject.itBehavesLike
 import spark.Service
 import java.net.URL
-import java.nio.file.Paths
 import java.util.concurrent.TimeUnit
 
 object DefaultLoadersSpec : SubjectSpek<DefaultLoaders>({
@@ -179,6 +177,7 @@ object DefaultLoadersSpec : SubjectSpek<DefaultLoaders>({
                 assertThat(newValue, equalTo("newValue"))
             }
         }
+        /* TODO: Move to konf-jgit
         on("load from git repository") {
             createTempDir().let { dir ->
                 Git.init().apply {
@@ -239,6 +238,7 @@ object DefaultLoadersSpec : SubjectSpek<DefaultLoaders>({
                 }
             }
         }
+        */
     }
 })
